requirements.txt
annotated-types==0.6.0
anyio==4.3.0
beautifulsoup4==4.12.3
black==22.1.0
blue==0.9.1
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
deepl==1.17.0
distro==1.9.0
flake8==4.0.1
google==3.0.0
google-api-core==2.18.0
google-auth==2.29.0
google-cloud-core==2.4.1
google-cloud-translate==3.15.3
googleapis-common-protos==1.63.0
grpcio==1.62.2
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
joblib==1.4.0
lxml==5.2.1
mccabe==0.6.1
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.23.6
pathspec==0.12.1
platformdirs==4.2.1
portalocker==2.8.2
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.8.0
pydantic==2.7.1
pydantic_core==2.18.2
pyflakes==2.4.0
python-dotenv==1.0.1
regex==2024.4.16
requests==2.31.0
rsa==4.9
sacrebleu==2.4.2
sacremoses==0.1.1
sniffio==1.3.1
soupsieve==2.5
tabulate==0.9.0
tomli==2.0.1
tqdm==4.66.2
typing_extensions==4.11.0
urllib3==2.2.1
